Added more type saftey for API .generate arguments.
